Make scio-avro compile
