+++ scio-avro/src/main/scala-2/com/spotify/scio/avro/types/SchemaProvider.scala
